requirements.txt
altair==5.5.0
attrs==25.1.0
Automat==24.8.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.17.0
gitdb==4.0.12
GitPython==3.1.44
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
MarkupSafe==3.0.2
narwhals==1.33.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==11.1.0
Protego==0.4.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
queuelib==1.7.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rpds-py==0.24.0
schedule==1.2.2
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
tenacity==9.1.2
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
w3lib==2.3.1
zope.interface==7.2
